Flip state and details in Discord status
Currently we have this implemented inverse. Training should be "details"
as it is more detailed description of what player is currently doing,
and "state" should be "In Game", "In Menu", "Exploring", "Raiding" etc,
as that describes best what is player currently doing (also, e.g for
discord party integration, it is appended after state, so there always
MUST be some state, and this PR ensures that).

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>